extract call date range viewer widget and improve detailed call stat
