+++ lib/widgets/call_date_range_viewer.dart
@@ -1,0 +1,111 @@
+import 'package:flutter/material.dart';
+
+class CallDateRangeViewer extends StatelessWidget {
+  final String initialDate, finalDate;
+
+  const CallDateRangeViewer({
+    super.key,
+    required this.initialDate,
+    required this.finalDate,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      padding: const EdgeInsets.symmetric(vertical: 10.0, horizontal: 8.0),
+      margin: const EdgeInsets.symmetric(vertical: 0.0, horizontal: 0.0),
+      decoration: BoxDecoration(
+        color: Colors.grey[200],
+        // border: Border.all(color: Colors.black),
+        borderRadius: const BorderRadius.all(
+          Radius.circular(10.0),
+        ),
+      ),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.center,
+        children: [
+          // Title
+          const Row(
+            mainAxisAlignment: MainAxisAlignment.start,
+            children: [
+              Padding(
+                padding: EdgeInsets.only(left: 4.0),
+                child: Icon(
+                  Icons.calendar_month_outlined,
+                  size: 17.0,
+                ),
+              ),
+              Padding(
+                padding: EdgeInsets.only(left: 5.0),
+                child: Text(
+                  "Date Range",
+                  style: TextStyle(
+                    fontSize: 16.0,
+                    fontWeight: FontWeight.w400,
+                  ),
+                ),
+              ),
+            ],
+          ),
+          const SizedBox(height: 8.0),
+          // Values
+          Container(
+            padding:
+                const EdgeInsets.symmetric(vertical: 10.0, horizontal: 10.0),
+            margin: const EdgeInsets.symmetric(vertical: 0.0, horizontal: 0.0),
+            decoration: BoxDecoration(
+              color: Colors.grey[300],
+              // border: Border.all(color: Colors.black),
+              borderRadius: const BorderRadius.all(
+                Radius.circular(10.0),
+              ),
+            ),
+            child: Row(
+              mainAxisAlignment: MainAxisAlignment.spaceAround,
+              children: [
+                Column(
+                  children: [
+                    Text(
+                      initialDate,
+                      style: const TextStyle(
+                        fontSize: 17.0,
+                        fontWeight: FontWeight.w500,
+                      ),
+                    ),
+                    const SizedBox(height: 2.0),
+                    Text(
+                      "from",
+                      style: TextStyle(
+                        fontSize: 14.0,
+                        color: Colors.grey[700],
+                      ),
+                    ),
+                  ],
+                ),
+                Column(
+                  children: [
+                    Text(
+                      finalDate,
+                      style: const TextStyle(
+                        fontSize: 17.0,
+                        fontWeight: FontWeight.w500,
+                      ),
+                    ),
+                    const SizedBox(height: 2.0),
+                    Text(
+                      "to",
+                      style: TextStyle(
+                        fontSize: 14.0,
+                        color: Colors.grey[700],
+                      ),
+                    ),
+                  ],
+                ),
+              ],
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
